+++ rolldown.config.ts
@@ -1,0 +1,42 @@
+import { join } from "node:path"
+import { defineConfig, OutputOptions } from "rolldown"
+import unpluginDts from "unplugin-dts"
+import unpluginPackage, {
+  defaultFilesToCopy,
+  defaultManifestOverride,
+} from "./src/index"
+
+const root = import.meta.dirname
+const outdir = join(root, "out")
+const buildCommon: OutputOptions = {
+  exports: "named",
+  minify: true,
+}
+
+export default defineConfig({
+  plugins: [
+    // Don't bundle dts because the unplugin export type might hang.
+    unpluginDts.rolldown(),
+    unpluginPackage.rolldown({
+      root,
+      outdir,
+      emptyOutdir: true,
+      copyFiles: [...defaultFilesToCopy, "README.zh.md", "LICENSE-mulan"],
+      manifestOverride(raw) {
+        const manifest = defaultManifestOverride(raw)
+        manifest["private"] = undefined
+        return manifest
+      },
+    }),
+  ],
+  external(id, _parentId, isResolved) {
+    if (isResolved) return false
+    if (id.startsWith("node:")) return true
+    return ["unplugin"].includes(id)
+  },
+  input: join(root, "src", "index.ts"),
+  output: [
+    { file: join(outdir, "index.js"), format: "esm", ...buildCommon },
+    { file: join(outdir, "index.cjs"), format: "commonjs", ...buildCommon },
+  ],
+})